consumer/requirements.txt
aiokafka==0.8.1
kafka-python==2.0.2
httpx==0.27.0
kafka-python
requests